requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.1.8
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.2.0
pamqp==3.3.0
propcache==0.3.1
psycopg==3.2.6
psycopg-binary==3.2.6
pyasn1==0.4.8
pydantic==2.11.0
pydantic-settings==2.8.1
pydantic_core==2.33.0
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlalchemy-cockroachdb==2.0.2
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
wheel==0.45.1
yarl==1.18.3
